pyproject.toml
[project]
name = "crawl-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio>=1.73.0",
    "grpcio-tools>=1.73.0",
    "protobuf>=6.31.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.0",
    "selenium>=4.33.0",
    "undetected-chromedriver>=3.5.5",
    "webdriver-manager>=4.0.2",
]
[tool.setuptools]
packages = ["config", "models", "proto", "repositories", "scraper", "services"]